Check session created
As part of our develpment we want to check the session
this can be done by the cookie that is stored on the request
so, we can check at any time if our request have that cookie.
We must get the value, and check if that value is present on our database.